requirements.txt
django==3.0.8
djangorestframework==3.11.0
django-cors-headers==3.7.0
Pillow==8.1.2
psycopg2-binary==2.8.6
gunicorn==20.1.0
drf-yasg==1.20.0
requests==2.26.0
